+++ src/Typin/Examples/BlazorExample/CLI/Commands/WebHostCommand.cs
@@ -24,8 +24,8 @@
 
         public async ValueTask ExecuteAsync(IConsole console)
         {
-            if (_cliContext.ModeSwitcher.Current == CliModes.Interactive)
-                throw new CommandException(Message, exitCode: 0, showHelp: false);
+            //if (_cliContext.ModeSwitcher.Current == CliModes.Interactive)
+            //    throw new CommandException(Message, exitCode: 0, showHelp: false);
 
             await _webHostRunnerService.RunAsync(console.GetCancellationToken());
         }