+++ src/Typin/Examples/InteractiveModeExample/Program.cs
@@ -7,6 +7,7 @@
     using Microsoft.Extensions.DependencyInjection;
     using Typin;
     using Typin.Directives;
+    using Typin.Modes;
 
     public static class Program
     {
@@ -26,7 +27,8 @@
                 .UseMiddleware<ExitCodeMiddleware>()
                 .UseMiddleware<ExecutionTimingMiddleware>()
                 .UseMiddleware<ExecutionLogMiddleware>()
-                .UseInteractiveMode()
+                .RegisterMode<DirectMode>()
+                .RegisterMode<InteractiveMode>()
                 .UseStartupMessage("{title} CLI {version} {{title}} {executable} {{{description}}} {test}")
                 .Build()
                 .RunAsync();