+++ src/Typin/Typin/Internal/Exceptions/EndUserExceptions.cs
@@ -5,9 +5,10 @@
     using Typin.Schemas;
 
     /// <summary>
-    /// End-user-facing exceptions. Avoid internal details and fix recommendations here
+    /// Interactive mode end-user-facing exceptions. Avoid internal details and fix recommendations here
     /// </summary>
-    internal static class EndUserExceptions
+    internal static class InteractiveModeEndUserExceptions
+
     {
         internal static TypinException InteractiveOnlyCommandButThisIsDirectApplication(CommandSchema command)
         {